backend/requirements.txt
fastapi==0.129.0
uvicorn[standard]==0.41.0
motor==3.7.1
pymongo==4.16.0
python-jose[cryptography]==3.5.0
bcrypt==5.0.0
python-multipart==0.0.22
pydantic[email]==2.12.5
pydantic-settings==2.13.0
python-dotenv==1.2.1
beanie==2.0.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-oauthlib==1.2.4
google-auth-httplib2==0.3.0
razorpay==2.0.0
minio==7.2.20
httpx==0.28.1
